Use new AST base parser by default
