+++ src/jsonpath.rs
@@ -4,8 +4,9 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+use crate::ast;
 use crate::parser;
-use crate::path::Path;
+use serde_json::Value;
 
 #[derive(Debug)]
 pub struct SyntaxError {
@@ -18,6 +19,19 @@ impl std::fmt::Display for SyntaxError {
     }
 }
 
-pub fn parse(selector: &str) -> Result<impl Path, SyntaxError> {
-    parser::parse(selector).map_err(|m| SyntaxError { message: m })
+pub enum FindError {
+    // no errors yet
+}
+
+pub fn parse(selector: &str) -> Result<Path, SyntaxError> {
+    let p = parser::parse(selector).map_err(|m| SyntaxError { message: m })?;
+    Ok(Path(p))
+}
+
+pub struct Path(ast::Path);
+
+impl Path {
+    pub fn find<'a>(&'a self, document: &'a Value) -> Result<Vec<&'a Value>, FindError> {
+        Ok(self.0.find(document).collect())
+    }
 }
